requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coverage==5.5
coveralls==3.0.1
Django==3.2
django-cors-headers==3.7.0
django-environ==0.4.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
docopt==0.6.2
idna==2.10
marshmallow==3.11.1
mypy-extensions==0.4.3
mysqlclient==2.0.3
pathspec==0.8.1
PyJWT==2.1.0
pytz==2021.1
regex==2021.4.4
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.26.4
